+++ packages/spl-token-bonding/src/curves.ts
@@ -14,48 +14,124 @@ export function fromCurve(
   curve: any,
   baseStorage: AccountInfo,
   baseMint: MintInfo,
-  targetMint: MintInfo
+  targetMint: MintInfo,
+  goLiveUnixTime: number
 ): IPricingCurve {
   switch (Object.keys(curve.definition)[0]) {
     case "timeV0":
-      const curv = curve.definition.timeV0.curves[0].curve.exponentialCurveV0;
-      return new ExponentialCurve(
-        curv as ExponentialCurveV0,
+      return new TimeCurve({
+        curve,
         baseStorage,
         baseMint,
-        targetMint
-      );
+        targetMint,
+        goLiveUnixTime
+      })
+
   }
 
   throw new Error("Curve not found");
 }
 
 export interface IPricingCurve {
-  current(): number;
+  current(unixTime?: number): number;
   locked(): number;
   sellTargetAmount(
     targetAmountNum: number,
     baseRoyaltiesPercent: number,
-    targetRoyaltiesPercent: number
+    targetRoyaltiesPercent: number,
+    unixTime?: number
   ): number;
   buyTargetAmount(
     targetAmountNum: number,
     baseRoyaltiesPercent: number,
-    targetRoyaltiesPercent: number
+    targetRoyaltiesPercent: number,
+    unixTime?: number
   ): number;
   buyWithBaseAmount(
     baseAmountNum: number,
     baseRoyaltiesPercent: number,
-    targetRoyaltiesPercent: number
+    targetRoyaltiesPercent: number,
+    unixTime?: number
   ): number;
   buyWithBaseRootEstimates(
     baseAmountNum: number,
-    baseRoyaltiesPercent: number
+    baseRoyaltiesPercent: number,
+    unixTime?: number
   ): number[];
   buyTargetAmountRootEstimates(
     targetAmountNum: number,
-    targetRoyaltiesPercent: number
+    targetRoyaltiesPercent: number,
+    unixTime?: number
   ): number[];
+}
+
+type TimeCurveArgs = {
+  curve: any,
+  baseStorage: AccountInfo;
+  baseMint: MintInfo;
+  targetMint: MintInfo;
+  goLiveUnixTime: number;
+}
+
+export class TimeCurve implements IPricingCurve {
+  curve: any;
+  baseStorage: AccountInfo;
+  baseMint: MintInfo;
+  targetMint: MintInfo;
+  goLiveUnixTime: number;
+
+  constructor({
+    curve,
+    baseStorage,
+    baseMint,
+    targetMint,
+    goLiveUnixTime
+  }: TimeCurveArgs) {
+    this.curve = curve;
+    this.baseStorage = baseStorage;
+    this.baseMint = baseMint;
+    this.targetMint = targetMint;
+    this.goLiveUnixTime = goLiveUnixTime;
+  }
+
+  currentCurve(unixTime: number = (new Date().valueOf() / 1000)): IPricingCurve {
+    let subCurve; 
+    if (unixTime < this.goLiveUnixTime) {
+      subCurve = this.curve.definition.timeV0.curves[0].curve.exponentialCurveV0
+    } else {
+      subCurve = [...this.curve.definition.timeV0.curves].reverse().find((c: any) => unixTime >= this.goLiveUnixTime + c.offset.toNumber()).curve.exponentialCurveV0
+    }
+
+    return new ExponentialCurve(
+      subCurve as ExponentialCurveV0,
+      this.baseStorage,
+      this.baseMint,
+      this.targetMint
+    );
+  }
+
+  current(unixTime?: number): number {
+    return this.currentCurve(unixTime).current(unixTime);
+  }
+
+  locked(): number {
+    return this.currentCurve().locked();
+  }
+  sellTargetAmount(targetAmountNum: number, baseRoyaltiesPercent: number, targetRoyaltiesPercent: number, unixTime?: number): number {
+    return this.currentCurve(unixTime).sellTargetAmount(targetAmountNum, baseRoyaltiesPercent, targetRoyaltiesPercent);
+  }
+  buyTargetAmount(targetAmountNum: number, baseRoyaltiesPercent: number, targetRoyaltiesPercent: number, unixTime?: number): number {
+    return this.currentCurve(unixTime).buyTargetAmount(targetAmountNum, baseRoyaltiesPercent, targetRoyaltiesPercent);
+  }
+  buyWithBaseAmount(baseAmountNum: number, baseRoyaltiesPercent: number, targetRoyaltiesPercent: number, unixTime?: number): number {
+    return this.currentCurve(unixTime).buyWithBaseAmount(baseAmountNum, baseRoyaltiesPercent, targetRoyaltiesPercent);
+  }
+  buyWithBaseRootEstimates(baseAmountNum: number, baseRoyaltiesPercent: number, unixTime?: number): number[] {
+    return this.currentCurve(unixTime).buyWithBaseRootEstimates(baseAmountNum, baseRoyaltiesPercent);
+  }
+  buyTargetAmountRootEstimates(targetAmountNum: number, targetRoyaltiesPercent: number, unixTime?: number): number[] {
+    return this.currentCurve(unixTime).buyTargetAmountRootEstimates(targetAmountNum, targetRoyaltiesPercent);
+  }
 }
 
 export class ExponentialCurve implements IPricingCurve {
@@ -139,10 +215,12 @@ export class ExponentialCurve implements IPricingCurve {
     baseRoyaltiesPercent: number,
     targetRoyaltiesPercent: number
   ): number {
-    // Calculate with the actual target amount they will need to get the target amount after royalties
-    const dS = targetAmountNum * (1 / (1 - asDecimal(targetRoyaltiesPercent)));
     const R = amountAsNum(this.baseStorage.amount, this.baseMint);
     const S = supplyAsNum(this.targetMint);
+    
+    // Calculate with the actual target amount they will need to get the target amount after royalties
+    const dS = targetAmountNum * (1 / (1 - asDecimal(targetRoyaltiesPercent)));
+
     if (R == 0 || S == 0) {
       // b dS + (c dS^(1 + k))/(1 + k)
       return (